- Manifest Internet Permission Added. - Binding added to gradle. - CoinDetailDto.kt, Constants.kt, CurrentPriceModel.kt, CoinRepository.kt, CoinApi.kt updated. - Load Image function added for binding.

+++ app/src/main/java/com/tracker/coin/data/remote/dto/CurrentPriceModel.kt
@@ -1,3 +1,3 @@
 package com.tracker.coin.data.remote.dto
 
-data class CurrentPriceModel(val usd: Int, val eur: Int)
+data class CurrentPriceModel(val usd: Double, val eur: Double)
